Raspberry compile fixes.

+++ avGraph/graphEdit_nodeTypeSelect.cpp
@@ -86,7 +86,7 @@ static void measureCategoryButton(const char * name, int & sx, int & sy)
 	float textSy;
 	measureText(kCategoryFontSize, textSx, textSy, name);
 	
-	sx = int(std::ceilf(textSx)) + kCategoryPadding * 2;
+	sx = int(std::ceil(textSx)) + kCategoryPadding * 2;
 	sy = kCategoryHeight;
 }
